Update user section
