+++ src/i18n/en/meleeData.ts
@@ -2,7 +2,7 @@ interface MeleeData {
     [key: string]: string;
 }
 
-export const melee_en: MeleeData = {
+export const meleeEN: MeleeData = {
     '0000': 'None',
     '0100': 'Iron Sword',
     '0200': 'Iron Sword α',
