+++ src/lib/manual/Alternate.svelte
