+++ src/lib/manual/BottomNav.svelte
