+++ src/lib/manual/LandscapeMode.svelte
