+++ src/lib/manual/LangSelectArea.svelte
