+++ src/lib/manual/ManualFooter.svelte
