+++ src/lib/manual/ManualHeader.svelte
