+++ src/lib/manual/ManualMenu.svelte
