+++ src/routes/+error.svelte
@@ -2,7 +2,7 @@
     import type { LayoutData } from './$types';
     import { setLocale } from '$i18n/i18n-svelte';
     import LL from '$i18n/i18n-svelte';
-    import Alternate from '$lib/common/Alternate.svelte';
+    import Alternate from '$lib/manual/Alternate.svelte';
     import '$scss/global.scss';
 
     export let data: LayoutData;
