+++ static/sass/global/_variables.scss
@@ -21,4 +21,4 @@ $infoColorMap: (
 );
 
 // 武器種名
-$weaponNamesArray: GreatSword, HeavyBowgun, Hammer, Lance, Sword&Shield, LightBowgun, DualBlades, LongSword, HuntingHorn, Gunlance, Bow, Tonfa, SwitchAxeF, MagnetSpike;
+$weaponNamesArray: GreatSword, HeavyBowgun, Hammer, Lance, Sword&Shield, LightBowgun, DualBlades, LongSword, HuntingHorn, Gunlance, Bow, Tonfa, SwitchAxeF, MagnetSpike, new;
